// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any };
  /** A field whose value conforms with the standard mongodb object Id as described here: https://docs.mongodb.com/manual/reference/method/ObjectId/#ObjectId. Example: 5e5677d71bdc2ae76344968c */
  ObjectId: { input: any; output: any };
};
// ...
export type Badge = Node & {
  __typename?: 'Badge';
  /** The date the badge was earned. */
  dateAssigned: Maybe<Scalars['DateTime']['output']>;
  /** The description of the badge. */
  description: Maybe<Scalars['String']['output']>;
  /** The ID of the badge. */
  id: Scalars['ID']['output'];
  /** The image of the badge. */
  image: Scalars['String']['output'];
  /** Link to badge page on Hashnode. */
  infoURL: Maybe<Scalars['String']['output']>;
  /** The name of the badge. */
  name: Scalars['String']['output'];
  /** A flag to determine if the badge is hidden. */
  suppressed: Maybe<Scalars['Boolean']['output']>;
};

/**
 * Contains basic information about the beta feature.
 * A beta feature is a feature that is not yet released to all users.
 */
export type BetaFeature = Node & {
  __typename?: 'BetaFeature';
  /** The description of the beta feature. */
  description: Maybe<Scalars['String']['output']>;
  /** The date the beta feature was created. */
  enabled: Scalars['Boolean']['output'];
  /** The ID of the beta feature. */
  id: Scalars['ID']['output'];
  /** The key of the beta feature. */
  key: Scalars['String']['output'];
  /** The title of the beta feature. */
  title: Maybe<Scalars['String']['output']>;
  /** The url of the beta feature. */
  url: Maybe<Scalars['String']['output']>;
};
// ...
export type Connection = {
  /** A list of edges of items connection. */
  edges: Array<Edge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

export type Content = {
  __typename?: 'Content';
  /** The HTML version of the content. */
  html: Scalars['String']['output'];
  /** The Markdown version of the content. */
  markdown: Scalars['String']['output'];
  /** The text version from sanitized html content. HTML tags are stripped and only text is returned. */
  text: Scalars['String']['output'];
};
// ...
export type IUser = {
  /** Whether or not the user is an ambassador. */
  ambassador: Scalars['Boolean']['output'];
  /** The availability of the user based on tech stack and interests. Shown on the "I am available for" section in user's profile. */
  availableFor: Maybe<Scalars['String']['output']>;
  /** Returns a list of badges that the user has earned. Shown on blogs /badges page. Example - https://iamshadmirza.com/badges */
  badges: Array<Badge>;
  /** The bio of the user. Visible in about me section of the user's profile. */
  bio: Maybe<Content>;
  /** The date the user joined Hashnode. */
  dateJoined: Maybe<Scalars['DateTime']['output']>;
  /** Whether or not the user is deactivated. */
  deactivated: Scalars['Boolean']['output'];
  /** The users who are following this user */
  followers: UserConnection;
  /** The number of users that follow the requested user. Visible in the user's profile. */
  followersCount: Scalars['Int']['output'];
  /** The number of users that this user is following. Visible in the user's profile. */
  followingsCount: Scalars['Int']['output'];
  /** The users which this user is following */
  follows: UserConnection;
  /** The ID of the user. It can be used to identify the user. */
  id: Scalars['ID']['output'];
  /** The location of the user. */
  location: Maybe<Scalars['String']['output']>;
  /** The name of the user. */
  name: Scalars['String']['output'];
  /** Returns the list of posts the user has published. */
  posts: UserPostConnection;
  /** The URL to the profile picture of the user. */
  profilePicture: Maybe<Scalars['String']['output']>;
  /** Publications associated with the user. Includes personal and team publications. */
  publications: UserPublicationsConnection;
  /** The social media links of the user. Shown on the user's profile. */
  socialMediaLinks: Maybe<SocialMediaLinks>;
  /** The tagline of the user. Shown on the user's profile below the name. */
  tagline: Maybe<Scalars['String']['output']>;
  /** Returns a list of tags that the user follows. */
  tagsFollowing: Array<Tag>;
  /** The username of the user. It is unique and tied with user's profile URL. Example - https://hashnode.com/@username */
  username: Scalars['String']['output'];
};
// ...
export type MyUser = IUser &
  Node & {
    __typename?: 'MyUser';
    /**
     * Whether or not the user is an ambassador.
     * @deprecated Ambassadors program no longer active. Will be removed after 02/01/2024
     */
    ambassador: Scalars['Boolean']['output'];
    /** The availability of the user based on tech stack and interests. Shown on the "I am available for" section in user's profile. */
    availableFor: Maybe<Scalars['String']['output']>;
    /** Returns a list of badges that the user has earned. Shown on blogs /badges page. Example - https://iamshadmirza.com/badges */
    badges: Array<Badge>;
    /** A list of beta features that the user has access to. Only available to the authenticated user. */
    betaFeatures: Array<BetaFeature>;
    /** The bio of the user. Visible in about me section of the user's profile. */
    bio: Maybe<Content>;
    /** The date the user joined Hashnode. */
    dateJoined: Maybe<Scalars['DateTime']['output']>;
    /** Whether or not the user is deactivated. */
    deactivated: Scalars['Boolean']['output'];
    /** Email address of the user. Only available to the authenticated user. */
    email: Maybe<Scalars['String']['output']>;
    /** The users who are following this user */
    followers: UserConnection;
    /** The number of users that follow the requested user. Visible in the user's profile. */
    followersCount: Scalars['Int']['output'];
    /** The number of users that this user is following. Visible in the user's profile. */
    followingsCount: Scalars['Int']['output'];
    /** The users which this user is following */
    follows: UserConnection;
    /** The ID of the user. It can be used to identify the user. */
    id: Scalars['ID']['output'];
    /** The location of the user. */
    location: Maybe<Scalars['String']['output']>;
    /** The name of the user. */
    name: Scalars['String']['output'];
    /** Returns the list of posts the user has published. */
    posts: UserPostConnection;
    /** The URL to the profile picture of the user. */
    profilePicture: Maybe<Scalars['String']['output']>;
    provider: Maybe<Scalars['String']['output']>;
    /** Publications associated with the user. Includes personal and team publications. */
    publications: UserPublicationsConnection;
    /** The social media links of the user. Shown on the user's profile. */
    socialMediaLinks: Maybe<SocialMediaLinks>;
    /** The tagline of the user. Shown on the user's profile below the name. */
    tagline: Maybe<Scalars['String']['output']>;
    /** Returns a list of tags that the user follows. */
    tagsFollowing: Array<Tag>;
    /** Hashnode users are subscribed to a newsletter by default. This field can be used to unsubscribe from the newsletter. Only available to the authenticated user. */
    unsubscribeCode: Maybe<Scalars['String']['output']>;
    /** The username of the user. It is unique and tied with user's profile URL. Example - https://hashnode.com/@username */
    username: Scalars['String']['output'];
  };
// ...
export type Node = {
  /** The ID of the node. */
  id: Scalars['ID']['output'];
};

/** Contains information to help in pagination for page based pagination. */
export type OffsetPageInfo = {
  __typename?: 'OffsetPageInfo';
  /** Indicates if there are more pages. */
  hasNextPage: Maybe<Scalars['Boolean']['output']>;
  /** Indicates if there are previous pages */
  hasPreviousPage: Maybe<Scalars['Boolean']['output']>;
  /**
   * The page after the current page.
   * Use it to build page navigation
   */
  nextPage: Maybe<Scalars['Int']['output']>;
  /**
   * The page before the current page.
   * Use it to build page navigation
   */
  previousPage: Maybe<Scalars['Int']['output']>;
};
// ...
export type PageConnection = {
  /** A list of edges of items connection. */
  nodes: Array<Node>;
  /** Information to aid in pagination. */
  pageInfo: OffsetPageInfo;
};
// ...
export type SocialMediaLinks = {
  __typename?: 'SocialMediaLinks';
  /** The user's Facebook profile. */
  facebook: Maybe<Scalars['String']['output']>;
  /** The user's GitHub profile. */
  github: Maybe<Scalars['String']['output']>;
  /** The user's Instagram profile. */
  instagram: Maybe<Scalars['String']['output']>;
  /** The user's LinkedIn profile. */
  linkedin: Maybe<Scalars['String']['output']>;
  /** The user's StackOverflow profile. */
  stackoverflow: Maybe<Scalars['String']['output']>;
  /** The user's Twitter profile. */
  twitter: Maybe<Scalars['String']['output']>;
  /** The user's website. */
  website: Maybe<Scalars['String']['output']>;
  /** The user's YouTube profile. */
  youtube: Maybe<Scalars['String']['output']>;
};
// ...
export type Tag = ITag &
  Node & {
    __typename?: 'Tag';
    /** Total number of users following this tag. */
    followersCount: Scalars['Int']['output'];
    /** The ID of the tag. */
    id: Scalars['ID']['output'];
    /** Information about the tag. Contains markdown html and text version of the tag's info. */
    info: Maybe<Content>;
    /** The logo of the tag. Shown in tag page. */
    logo: Maybe<Scalars['String']['output']>;
    /** The name of the tag. Shown in tag page. */
    name: Scalars['String']['output'];
    /** Paginated list of posts published under this tag */
    posts: FeedPostConnection;
    /** Alltime usage count of this tag in posts. */
    postsCount: Scalars['Int']['output'];
    /** The slug of the tag. Used to access tags feed.  Example https://hashnode.com/n/graphql */
    slug: Scalars['String']['output'];
    /** The tagline of the tag. */
    tagline: Maybe<Scalars['String']['output']>;
  };
// ...
export type UserConnection = PageConnection & {
  __typename?: 'UserConnection';
  /** A list of users */
  nodes: Array<User>;
  /** Information for page based pagination in users connection. */
  pageInfo: OffsetPageInfo;
  /** The total number of documents in the connection. */
  totalDocuments: Scalars['Int']['output'];
};
// ...
export type UserPostConnection = PageConnection & {
  __typename?: 'UserPostConnection';
  /** A list of edges containing Post information */
  edges: Array<UserPostEdge>;
  /** A list of posts */
  nodes: Array<Post>;
  /** Information for page based pagination in Post connection. */
  pageInfo: OffsetPageInfo;
  /** The total number of documents in the connection. */
  totalDocuments: Scalars['Int']['output'];
};
// ...
export type UserPostEdge = {
  __typename?: 'UserPostEdge';
  /** Indicates weather the user is the author or co-author of the post. */
  authorType: PostAuthorType;
  /** The node holding the Post information. */
  node: Post;
};
// ...
export type UserPublicationsConnection = Connection & {
  __typename?: 'UserPublicationsConnection';
  /** A list of edges of publications connection. */
  edges: Array<UserPublicationsEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The total amount of publications taking into account the filter. */
  totalDocuments: Scalars['Int']['output'];
};

/** Filter to apply to the publications. */
export type UserPublicationsConnectionFilter = {
  /** Only return pro publications. */
  isPro: InputMaybe<Scalars['Boolean']['input']>;
  /** Only include publication in which the user has one of the provided roles. */
  roles: InputMaybe<Array<UserPublicationRole>>;
};

/** An edge that contains a node of type publication and cursor to the node. */
export type UserPublicationsEdge = Edge & {
  __typename?: 'UserPublicationsEdge';
  /** The cursor to the node. */
  cursor: Scalars['String']['output'];
  /** Node containing the publication. */
  node: Publication;
  /** The role of the user in the publication. */
  role: UserPublicationRole;
};